test: split PDF heading classification cases
